lots of rethinking

+++ spotify/playlist_generator.py
@@ -4,6 +4,9 @@ import spotipy
 import spotipy.util as util
 
 import os
+import json
+from json.decoder import JSONDecodeError
+from time import sleep
 
 os.environ["SPOTIPY_CLIENT_ID"] = SPOTIFY_API_ID
 os.environ["SPOTIPY_CLIENT_SECRET"] = SPOTIFY_API_SECRET
@@ -22,9 +25,17 @@ def main():
     sp = connect_to_spotify()
     print("Hello, World!")
 
-    tracks = load_track_list()
+    force_data = True
 
-    print(tracks)
+    tracks = load_track_list()
+    if tracks == False:
+        force_data = True
+        tracks = build_track_list()
+
+    data = analyze_track_list(tracks[:20])
+
+    for item in data:
+        print(item)
 
 
 def connect_to_spotify():
@@ -33,7 +44,8 @@ def connect_to_spotify():
     try:
         token = util.prompt_for_user_token(username, scope)
     except:
-        os.remove(f".cache-{username}")
+        if os.path.exists(f".cache-{username}"):
+            os.remove(f".cache-{username}")
         token = util.prompt_for_user_token(username, scope)
 
     spotifyObject = spotipy.Spotify(auth=token)
@@ -48,7 +60,7 @@ def load_track_list():
             data = json.load(file)
         return data
     except:
-        log("ERROR - No Save File Found")
+        log("ERROR - Failed to Load")
         return False
 
 
@@ -88,6 +100,48 @@ def save_track_data(_data):
     except:
         log("ERROR - Failed To Save")
         return False
+
+
+def build_track_list():
+    global sp
+    log("Building Track List")
+    total = 0
+    offset = 0
+    page_limit = 50
+    tracks = []
+    while True:
+        page = sp.current_user_saved_tracks(limit=page_limit, offset=offset)
+        if total == 0:
+            total = page["total"]
+        tracks += page["items"]
+        if offset + page_limit >= total:
+            leftovers = total - offset
+            tracks += sp.current_user_saved_tracks(limit=leftovers, offset=offset)[
+                "items"
+            ]
+            break
+        offset += page_limit
+        print(offset, "/", total, end="\r")
+        sleep(speed / 1000)
+    print(offset + leftovers, "/", total)
+    save_track_list(tracks)
+    return tracks
+
+
+def analyze_track_list(_tracks):
+    global sp
+    tracks = []
+    for track in _tracks:
+        # print("Artist: ", track["track"]["artists"][0]["name"])
+        tracks.append(
+            {
+                "artist": track["track"]["artists"][0]["name"],
+                "song": track["track"]["name"],
+                "id": track["track"]["id"],
+                "genre": sp.album(track["track"]["album"]["id"])["genres"],
+            }
+        )
+    return tracks
 
 
 def log(_msg):
